+++ src/main.py
@@ -1,14 +1,13 @@
 import argparse
-import json
 import random
 import requests
-import time
 import logging as log
-from random import randrange
 from logging.handlers import RotatingFileHandler
 
 from config import Settings, PixelFedBotException
 from dal import create_tables, load_followers
+from timelines import get_timeline_url, get_timeline
+from utils import random_time
 
 settings = Settings()
 handlers = [
@@ -22,37 +21,8 @@ handlers = [
 log.basicConfig(format='%(asctime)s | %(levelname)s | %(message)s', handlers=handlers, level=log.INFO)
 
 
-timeline_types = ['home', 'public', 'notifications', 'global']
-token = settings.token
-BASE_URL = 'https://pixelfed.social/'
-API_VERSION = 'api/v1/'
+timeline_types = ['home', 'public', 'notifications', 'global', 'tags']
 verify_cred_endpoint = 'accounts/verify_credentials'
-headers = {
-    "Authorization": f"Bearer {token}"
-}
-
-
-def random_time():
-    '''Use this to randomize actions'''
-    sleep_time = randrange(5, 120)
-    log.info(f'sleeping for {sleep_time} seconds...')
-    time.sleep(sleep_time)
-    return sleep_time
-
-
-def get_timeline(url: str, timeline_type: str = 'home', limit: int = 10) -> dict:
-    log.info(f'getting timeline {timeline_type} @ {url}')
-    params = {
-        "min_id": 1,
-        "limit": limit,
-    }
-    response = requests.get(url, headers=headers, params=params)
-    if response.status_code == 200:
-        log.info('Response successful')
-        return response.json()
-    else:
-        log.info(f"Failed to fetch data. Status code: {response.status_code}")
-        return {}
 
 
 def parse_timeline_for_favorites(data: list, limit: int = None) -> list:
@@ -77,8 +47,8 @@ def get_random_followers() -> list:
 
 
 def fave_post(status_id) -> int:
-    url = f'{BASE_URL}{API_VERSION}statuses/{status_id}/favourite'
-    response = requests.post(url, headers=headers)
+    url = f'{settings.base_url}{settings.api_version}statuses/{status_id}/favourite'
+    response = requests.post(url, headers=settings.headers)
 
     if response.status_code == 200:
         log.info(f'fave id: {status_id} request successful!')
@@ -102,10 +72,10 @@ def filter_notification_faves(data: list, limit: int = 5) -> list:
 
 
 def get_status_by_id(id: str, limit: int = 6, follower: str = None) -> dict:
-    url = f'{BASE_URL}{API_VERSION}accounts/{id}/statuses'
+    url = f'{settings.base_url}{settings.api_version}accounts/{id}/statuses'
     param = {'limit': str(limit)}
     log.info(f'getting timeline {follower or id} @ {url}')
-    response = requests.get(url, headers=headers, params=param)
+    response = requests.get(url, headers=settings.headers, params=param)
     return response.json()
 
 
@@ -118,33 +88,12 @@ def fave_unfaved(server_response: dict, limit: int = 6):
     return liked_count
 
 
-def write_to_json(data):
-    with open('response.json', 'w') as json_file:
-        json.dump(data, json_file, indent=4)
-
-
-def read_json(data):
-    with open(data, 'r') as json_file:
-        return json.load(json_file)
-
-
-def get_timeline_url(timeline_type: str) -> tuple:
-    timeline_base = f'{BASE_URL}{API_VERSION}timelines'
-    if timeline_type == 'global':
-        return (f'{timeline_base}/public?min_id=1&limit=6&_pe=1&remote=true', timeline_type)
-    if timeline_type == 'notifications':
-        return (f'{BASE_URL}{API_VERSION}{timeline_type}', timeline_type)
-    if timeline_type == 'followers':
-        return (f'{BASE_URL}{API_VERSION}accounts/{settings.account_id}/{timeline_type}?limit=50', timeline_type)
-    return (f'{timeline_base}/{timeline_type}', timeline_type)
-
-
 def is_like_per_session_fulfilled(like_count: int) -> bool:
     return like_count >= settings.likes_per_session
 
 
 def process_notification_timeline(url_args: tuple, like_count: int = 0) -> int:
-    server_response = get_timeline(url=url_args[0], timeline_type=url_args[1])
+    server_response = get_timeline(url=url_args[0], settings=settings, timeline_type=url_args[1])
     id_list = filter_notification_faves(server_response)
     for id in id_list:
         server_response = get_status_by_id(id, limit=6)
@@ -156,7 +105,7 @@ def process_notification_timeline(url_args: tuple, like_count: int = 0) -> int:
 
 
 def process_timeline(url_args: tuple) -> int:
-    server_response = get_timeline(url=url_args[0], timeline_type=url_args[1])
+    server_response = get_timeline(url=url_args[0], settings=settings, timeline_type=url_args[1])
     return fave_unfaved(server_response, limit=settings.likes_per_session)
 
 
@@ -190,11 +139,11 @@ def main():
         log.info('starting pixelfed bot')
         create_tables()
         settings.likes_per_session = args.limit or settings.likes_per_session
-        url_args = get_timeline_url(args.timeline_type)
+        url_args = get_timeline_url(args.timeline_type, settings)
         like_count = handle_timeline(url_args)
         log.info(f'first pass count: {like_count}')
         while not is_like_per_session_fulfilled(like_count):
-            log.info(f'Like count: {like_count} less than likes per session value: {settings.likes_per_session}')
+            log.info(f'Like count: {like_count}, per session value: {settings.likes_per_session}')
             random_time()
             new_likes = process_follower_timeline()
             like_count += new_likes
@@ -202,12 +151,11 @@ def main():
             if is_like_per_session_fulfilled(like_count):
                 break
             random.shuffle(timeline_types)
-            new_likes = handle_timeline(get_timeline_url(timeline_types[0]), like_count)
+            new_likes = handle_timeline(get_timeline_url(timeline_types[0], settings), like_count)
             like_count += new_likes
             log.info(f'Liked {new_likes} posts from {timeline_types[0]} timeline. Total likes: {like_count}')
-            # TODO add following list 
-            # TODO Add tag list to like : 	/api/v1/tags/nature
-        log.info(f'Reached total like count: {like_count} exceeding {settings.likes_per_session}')
+            # TODO add following list
+            log.info(f'Reached total like count: {like_count} exceeding {settings.likes_per_session}')
     except PixelFedBotException as ex:
         log.error(ex, exc_info=True)
